MongoDB authentication support.
With the new MONGO_USER and MONGO_PASS configuration settings one can
connect to a mongodb server that requires authentication.

The support is very basic: we always try to authenticate, and silently
proceed if it fails. If the database does require authentication,
queries will fail anyway, and an appropriate error will be returned.